Stop daily commits (remove lastUpdate checks)
Instead of constantly generating new salts, the salt file is fixed and can only be changed through workflows designed to run when dispatched manually.
This update will prevent ludicrous numbers of commits from piling up in the repository and also gets rid of the need for a waiting period as the word is generated entirely by the client.